minor formatting stuff

+++ telemetry_receiver/uart.h
@@ -5,5 +5,5 @@
 
 void InitUART();
 size_t UARTAvailable();
-uint8_t UARTReadBytes(uint8_t *data, const size_t size);
+int UARTReadBytes(uint8_t *data, const size_t size);
 uint8_t UARTReadOneByte();
